+++ IntuneWinPackager.Models/Enums/PackageCatalogSource.cs
@@ -1,0 +1,7 @@
+namespace IntuneWinPackager.Models.Enums;
+
+public enum PackageCatalogSource
+{
+    Winget = 0,
+    Chocolatey = 1
+}